Refactor squin noise tests into separate file
